default to notes in nav
