+++ app/(tabs)/_layout.tsx
@@ -31,15 +31,15 @@ export default function TabLayout() {
         <Tabs.Screen
           name="index"
           options={{
-            title: 'Home',
-            tabBarIcon: ({ color }) => <IconSymbol size={28} name="house.fill" color={color} />,
+            title: 'Notes',
+            tabBarIcon: ({ color }) => <IconSymbol size={28} name="note.text" color={color} />,
           }}
         />
         <Tabs.Screen
-          name="explore"
+          name="home"
           options={{
-            title: 'Notes',
-            tabBarIcon: ({ color }) => <IconSymbol size={28} name="note.text" color={color} />,
+            title: 'Home',
+            tabBarIcon: ({ color }) => <IconSymbol size={28} name="house.fill" color={color} />,
           }}
         />
       </Tabs>
